Constrain version of Cloudflare provider
Version 5 of the Cloudflare provider is a major rewrite with a number
of backwards incompatible changes. This patch release explicitly
notes that the provider should not be updated. A new version of this
module will be released shortly.

Migration guide: https://registry.terraform.io/providers/cloudflare/cloudflare/latest/docs/guides/version-5-upgrade